3/discord_bot/requirements.txt
requests==2.28.2
discord.py==2.2.2
